+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
 - [SA-1463] Return error when trying to create a new socket
+- [SA-2074] Add a ReadPacket function